update joints after culling
