+++ cocos/3d/assets/mesh.ts
@@ -26,6 +26,7 @@
 import { Asset } from '../../assets/asset';
 import { ccclass, property } from '../../core/data/class-decorator';
 import { Vec3 } from '../../core/value-types';
+import { vec3 } from '../../core/vmath';
 import { GFXBuffer } from '../../gfx/buffer';
 import { GFXAttributeName, GFXBufferUsageBit, GFXFormat, GFXFormatInfos, GFXFormatType, GFXMemoryUsageBit, GFXPrimitiveMode } from '../../gfx/define';
 import { GFXDevice } from '../../gfx/device';
@@ -33,7 +34,6 @@ import { IGFXAttribute } from '../../gfx/input-assembler';
 import { BufferBlob } from '../misc/buffer-blob';
 import { IBufferView } from './utils/buffer-view';
 import { postLoadMesh } from './utils/mesh-utils';
-import { vec3 } from '../../core/vmath';
 
 function getIndexStrideCtor (stride: number) {
     switch (stride) {
@@ -499,9 +499,6 @@ export class Mesh extends Asset {
 
             vbView.set(srcVBView);
 
-            // if (bundle.view.stride === dstBundle.view.stride) {
-            //     vbView.set(dstVBView, bundle.view.length);
-            // } else {
             srcAttrOffset = 0;
             for (const attr of bundle.attributes) {
                 dstVBOffset = 0;
@@ -525,7 +522,6 @@ export class Mesh extends Asset {
                 }
                 srcAttrOffset += GFXFormatInfos[attr.format].size;
             }
-            // }
 
             vertexBundles[i] = {
                 attributes: bundle.attributes,
