+++ cocos/renderer/models/skinning-model.ts
@@ -1,10 +1,11 @@
 // Copyright (c) 2017-2018 Xiamen Yaji Software Co., Ltd.
 
 import { Skeleton } from '../../3d/assets/skeleton';
+import { LCA } from '../../3d/misc/utils';
 import { Filter, PixelFormat, WrapMode } from '../../assets/asset-enum';
 import { Texture2D } from '../../assets/texture-2d';
 import { Mat4, Quat, Vec3 } from '../../core/value-types';
-import { mat4, quat, vec4 } from '../../core/vmath';
+import { mat4, quat, vec3 } from '../../core/vmath';
 import { GFXBuffer } from '../../gfx/buffer';
 import { GFXBufferUsageBit, GFXFormatInfos, GFXMemoryUsageBit } from '../../gfx/define';
 import { GFXDevice, GFXFeature } from '../../gfx/device';
@@ -46,15 +47,61 @@ const _jointsFormat = {
     [JointsMediumType.RGBA32F]: PixelFormat.RGBA32F,
 };
 
-const m4_1 = new Mat4();
+const v3_1 = new Vec3();
+const v3_2 = new Vec3();
 const qt_0 = new Quat();
 const qt_1 = new Quat();
+const qt_2 = new Quat();
+const m4_1 = new Mat4();
 const f4_1 = new Float32Array(4);
+
+export class Joint {
+    public node: Node;
+    public position: Vec3 = new Vec3();
+    public rotation: Quat = new Quat();
+    public scale: Vec3 = new Vec3(1, 1, 1);
+    public parent: Joint | null = null;
+    protected _lastUpdate = -1;
+    constructor (node: Node) { this.node = node; }
+    public update () {
+        const totalFrames = cc.director.getTotalFrames();
+        if (this._lastUpdate >= totalFrames) { return; }
+        this._lastUpdate = totalFrames;
+
+        const parent = this.parent;
+        if (parent) {
+            parent.update();
+            vec3.multiply(this.position, this.node.position, parent.scale);
+            vec3.transformQuat(this.position, this.position, parent.rotation);
+            vec3.add(this.position, this.position, parent.position);
+            quat.multiply(this.rotation, parent.rotation, this.node.rotation);
+            vec3.multiply(this.scale, parent.scale, this.node.scale);
+        }
+    }
+}
+
+class JointManager {
+    public static get (node: Node, root: Node) {
+        let joint = JointManager._joints.get(node);
+        if (joint) { return joint; }
+        joint = new Joint(node);
+        if (node !== root && node.parent) { joint.parent = JointManager.get(node.parent, root); }
+        JointManager._joints.set(node, joint);
+        return joint;
+    }
+    protected static _joints: Map<Node, Joint> = new Map();
+}
 
 export class SkinningModel extends Model {
     // change here and cc-skinning.inc to use other skinning algorithms
     public updateJointData = this.updateJointDataDQS;
 
+    get joints () {
+        return this._joints;
+    }
+
+    private _skeleton: Skeleton | null = null;
+    private _joints: Joint[] = [];
     private _jointsMedium: IJointsInfo | null = null;
 
     constructor (scene: RenderScene, node: Node) {
@@ -65,6 +112,7 @@ export class SkinningModel extends Model {
 
     public bindSkeleton (skeleton: Skeleton | null) {
         this._destroyJointsMedium();
+        this._skeleton = skeleton;
         if (!skeleton) { return; }
         const type = selectJointsMediumType(this._device, skeleton.joints.length);
         const format = _jointsFormat[type];
@@ -104,6 +152,41 @@ export class SkinningModel extends Model {
         }
     }
 
+    public updateUBOs () {
+        if (!super.updateUBOs() || !this._skeleton) { return false; }
+
+        const len = this._joints.length;
+        for (let i = 0; i < len; ++i) {
+            const cur = this._joints[i]; cur.update();
+            const bindpose = this._skeleton.bindposes[i];
+
+            vec3.multiply(v3_1, bindpose.position, cur.scale);
+            vec3.transformQuat(v3_1, v3_1, cur.rotation);
+            vec3.add(v3_1, v3_1, cur.position);
+            quat.multiply(qt_1, cur.rotation, bindpose.rotation);
+            vec3.multiply(v3_2, cur.scale, bindpose.scale);
+
+            this.updateJointData(i, v3_1, qt_1, v3_2, i === 0);
+        }
+
+        this.commitJointData();
+        return true;
+    }
+
+    public resetSkinningTarget (skinningRoot: Node) {
+        const root = LCA(this.node, skinningRoot);
+        if (!root || !this._skeleton) { return; }
+        this._joints.length = 0;
+        this._skeleton.joints.forEach((path) => {
+            const targetNode = skinningRoot.getChildByPath(path);
+            if (!targetNode) {
+                console.warn(`Skinning target ${path} not found in scene graph.`);
+                return;
+            }
+            this._joints.push(JointManager.get(targetNode, root));
+        });
+    }
+
     // Linear Blending Skinning
     protected updateJointDataLBS (idx: number, pos: Vec3, rot: Quat, scale: Vec3) {
         if (!this._jointsMedium) { return; }
@@ -133,17 +216,17 @@ export class SkinningModel extends Model {
         if (first) { quat.copy(qt_0, rot); }
         else if (quat.dot(qt_0, rot) < 0) { quat.scale(rot, rot, -1); }
         // conversion
-        quat.set(qt_1, pos.x, pos.y, pos.z, 0);
-        quat.scale(qt_1, quat.multiply(qt_1, qt_1, rot), 0.5);
+        quat.set(qt_2, pos.x, pos.y, pos.z, 0);
+        quat.scale(qt_2, quat.multiply(qt_2, qt_2, rot), 0.5);
         // upload
         out[base + 0] = rot.x;
         out[base + 1] = rot.y;
         out[base + 2] = rot.z;
         out[base + 3] = rot.w;
-        out[base + 4] = qt_1.x;
-        out[base + 5] = qt_1.y;
-        out[base + 6] = qt_1.z;
-        out[base + 7] = qt_1.w;
+        out[base + 4] = qt_2.x;
+        out[base + 5] = qt_2.y;
+        out[base + 6] = qt_2.z;
+        out[base + 7] = qt_2.w;
         out[base + 8] = scale.x;
         out[base + 9] = scale.y;
         out[base + 10] = scale.z;
